requirements.txt
aiogram==3.4.1
python-dotenv
cachetools
sqlalchemy==2.0
pandas==2.2.1